combine very similar files
