+++ plottr/utils/__init__.py
@@ -1,1 +1,2 @@
-from .data import *
+from .misc import *
+from .num import *
